Build static MUSL target with vendored openssl
Rust is capable of building binaries which carry their own MUSL
libc internally for operation on almost any userspace provided that
the kernel ABI (signals and such) is compatible with the platform
target selected.
This project runs into OpenSSL dependency problems when building a
MUSL target as the host libraries are generally built for GNU.
Address the concern by providing a vendored openssl, which permits
building for the target via `cargo build --features vendored` with
the relevant release and target/toolchain invocation.

Testing:
  This patch is applied to our internal PKGBUILD, producing working
binaries.